Fixed: Decimal Values in Dialog Boxes

+++ JaVisteAMinhaMedia/app/src/main/java/pt/isec/gps1819g11/javisteaminhamedia/DialogInputs/UpdateGradesDialog.java
@@ -77,7 +77,7 @@ public class UpdateGradesDialog extends DialogFragment implements View.OnClickLi
 
                 String value = String.valueOf(inputGrade.getText());
                 if(!value.isEmpty()){
-                   float grade = Integer.parseInt(value);
+                   float grade = Float.parseFloat(value);
                     if(grade <= 20 && grade > 9.5 ){
                         try {
                             student.setGrade(gradeName, grade);
